--- app/kitchen_api/usecase/create_ticket/ticket_sevice.go
@@ -1,19 +1,0 @@
-package create_ticket
-
-import (
-	"github.com/tkame123/ddd-sample/app/kitchen_api/domain/port/domain_event"
-	"github.com/tkame123/ddd-sample/app/kitchen_api/domain/port/repository"
-	"github.com/tkame123/ddd-sample/app/kitchen_api/domain/port/service"
-)
-
-type CreatTicketService struct {
-	rep repository.Repository
-	pub domain_event.Publisher
-}
-
-func NewService(rep repository.Repository, pub domain_event.Publisher) service.CreateTicket {
-	return &CreatTicketService{
-		rep: rep,
-		pub: pub,
-	}
-}
